links on lot info pane now link directly to zoning website
