+++ app/controllers/lot.js
@@ -1,6 +1,18 @@
 import Ember from 'ember';
 import { computed } from 'ember-decorators/object'; // eslint-disable-line
 import Bookmarkable from '../mixins/bookmarkable';
+import carto from '../utils/carto';
+
+import specialPurposeCrosswalk from '../utils/special-purpose-crosswalk';
+
+const SQL = function(table, geometry) {
+  return `SELECT * FROM ${table} 
+          WHERE 
+            ST_Intersects(
+              ST_SetSRID(
+                ST_GeomFromGeoJSON('${JSON.stringify(geometry)}'), 4326), 
+                ${table}.the_geom);`;
+};
 
 export default Ember.Controller.extend(Bookmarkable, {
 
@@ -8,4 +20,25 @@ export default Ember.Controller.extend(Bookmarkable, {
   paddedZonemap(zonemap) {
     return (`0${zonemap}`).slice(-3);
   },
+
+  @computed('lot.geometry')
+  parentSpecialPurposeDistricts(geometry) {
+    console.log(geometry);
+    return carto.SQL(SQL('special_purpose_districts_v201802', geometry))
+      .then(response => response.map(
+        (item) => {
+          const [, [anchorName, boroName]] = specialPurposeCrosswalk
+            .find(([dist]) => dist === item.sdname);
+
+          return {
+            id: item.cartodb_id,
+            label: item.sdlbl.toUpperCase(),
+            name: item.sdname,
+            anchorName,
+            boroName,
+          };
+        }),
+      );
+  },
+
 });
