+++ app/templates/lot.hbs
@@ -30,29 +30,10 @@
         {{#if lot.zonedist2}}<li><a target="_blank" href="https://www1.nyc.gov/site/planning/zoning/districts-tools/{{primaryzone2}}.page" class="button">{{fa-icon 'external-link'}} {{lot.zonedist2}}</a></li>{{/if}}
         {{#if lot.zonedist3}}<li><a target="_blank" href="https://www1.nyc.gov/site/planning/zoning/districts-tools/{{primaryzone3}}.page" class="button">{{fa-icon 'external-link'}} {{lot.zonedist3}}</a></li>{{/if}}
         {{#if lot.zonedist4}}<li><a target="_blank" href="https://www1.nyc.gov/site/planning/zoning/districts-tools/{{primaryzone4}}.page" class="button">{{fa-icon 'external-link'}} {{lot.zonedist4}}</a></li>{{/if}}
-        {{#if lot.spdist1}}
-          <li>
-            {{link-to-intersecting (concat 'Special District: ' lot.spdist1) 'special-purpose-district' 'special_purpose_districts_v201802' lot.geometry classNames="button"}}
-          </li>
-        {{/if}}
-
-        {{#if lot.spdist2}}
-          <li>
-            {{link-to-intersecting (concat 'Special District: ' lot.spdist2) 'special-purpose-district' 'special_purpose_districts_v201802' lot.geometry classNames="button"}}
-          </li>
-        {{/if}}
-
-        {{#if lot.spdist3}}
-          <li>
-            {{link-to-intersecting (concat 'Special District: ' lot.spdist3) 'special-purpose-district' 'special_purpose_districts_v201802' lot.geometry classNames="button"}}
-          </li>
-        {{/if}}
-
-        {{#if lot.spdist4}}
-          <li>
-            {{link-to-intersecting (concat 'Special District: ' lot.spdist4) 'special-purpose-district' 'special_purpose_districts_v201802' lot.geometry classNames="button"}}
-          </li>
-        {{/if}}
+        
+        {{#each (await parentSpecialPurposeDistricts) as |specialDistrict|}}
+              <li><a target="_blank" href="https://www1.nyc.gov/site/planning/zoning/districts-tools/special-purpose-districts-{{specialDistrict.boroName}}.page#{{specialDistrict.anchorName}}" class="button">{{fa-icon 'external-link'}} {{specialDistrict.label}}</a></li>
+        {{/each}}
 
         {{#if lot.overlay1}}
           <li>
