+++ app/utils/special-purpose-crosswalk.js
@@ -1,0 +1,120 @@
+const specialPurposeCrosswalk = [
+  ['Special City Island District', ['city_island', 'bronx']],
+  ['Special Fort Totten Natural Area District-4', ['natural_area', 'bronx']],
+  ['Special Natural Area District', ['natural_area', 'citywide']],
+  ['Special Hudson River Park District', ['hudson_river_park', 'manhattan']],
+  ['Special Mixed Use District (MX-10)', ['mixed_use', 'citywide']],
+  ['Special Southern Hunters Point District', ['hunters_point', 'queens']],
+  ['Special Scenic View District', ['scenic_view', 'citywide']],
+  ['Special Southern Roosevelt Island District', ['sri', 'manhattan']],
+  ['Special Mixed Use District (MX-8)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-8)', ['mixed_use', 'citywide']],
+  ['Special Hunts Point District', ['hunts_point', 'bronx']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special 125th Street District', ['125th', 'manhattan']],
+  ['Special Planned Community Preservation District', ['planned_community', 'citywide']],
+  ['Special Planned Community Preservation District', ['planned_community', 'citywide']],
+  ['Special Coney Island Mixed Use District', ['coney_island', 'brooklyn']],
+  ['Special Hudson River Park District', ['hudson_river_park', 'manhattan']],
+  ['Special Garment Center District', ['garment_center', 'manhattan']],
+  ['Special Little Italy District', ['little_italy', 'manhattan']],
+  ['Special Coastal Risk District', ['special_coastal_risk', 'citywide']],
+  ['Special Mixed Use District (MX-11)', ['mixed_use', 'citywide']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Enhanced Commercial District 4', ['special_enhanced', 'citywide']],
+  ['Special Hudson Square District', ['hudson_square', 'manhattan']],
+  ['Special Enhanced Commercial District- 2', ['special_enhanced', 'citywide']],
+  ['Special Enhanced Commercial District- 2', ['special_enhanced', 'citywide']],
+  ['Special Battery Park City District', ['battery_park', 'manhattan']],
+  ['Special Limited Commercial District', ['limited_commercial', 'citywide']],
+  ['Special Lincoln Square District', ['lincoln_square', 'manhattan']],
+  ['Special Manhattanville Mixed Use District', ['manhattanville', 'manhattan']],
+  ['Special Mixed Use District (MX-6)', ['mixed_use', 'citywide']],
+  ['Special Natural Area District', ['natural_area', 'citywide']],
+  ['Special Union Square District', ['union_square', 'manhattan']],
+  ['Special West Chelsea District', ['west_chelsea', 'manhattan']],
+  ['Special Willets Point District', ['willets_point', 'queens']],
+  ['Special Clinton District', ['clinton', 'manhattan']],
+  ['Special Mixed Use District (MX-12)', ['mixed_use', 'citywide']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Enhanced Commercial District- 2', ['special_enhanced', 'citywide']],
+  ['Special Enhanced Commercial District- 2', ['special_enhanced', 'citywide']],
+  ['Special Enhanced Commercial District 5', ['special_enhanced', 'citywide']],
+  ['Special Mixed Use District (MX-13)', ['mixed_use', 'citywide']],
+  ['Special Planned Community Preservation District', ['planned_community', 'citywide']],
+  ['Special Planned Community Preservation District', ['planned_community', 'citywide']],
+  ['Special Mixed Use District (MX-8)', ['mixed_use', 'citywide']],
+  ['Special Coastal Risk District', ['special_coastal_risk', 'citywide']],
+  ['Special Mixed Use District (MX-16)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-2)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-14)', ['mixed_use', 'citywide']],
+  ['Special Coney Island District', ['coney_island', 'brooklyn']],
+  ['Special Mixed Use District (MX-7)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-9)', ['mixed_use', 'citywide']],
+  ['Special Natural Area District', ['natural_area', 'bronx']],
+  ['Special Mixed Use District (MX-1)', ['mixed_use', 'citywide']],
+  ['Special Grand Concourse Preservation District', ['grand_concourse', 'bronx']],
+  ['Special Harlem River Waterfront District', ['harlem_river', 'bronx']],
+  ['Special Hillsides Preservation District', ['hillsides', 'staten-island']],
+  ['Special Hudson Yards District', ['hudson_yards', 'manhattan']],
+  ['Special Madison Avenue Preservation District', ['madison', 'citywide']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Mixed Use District (MX-4)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-5)', ['mixed_use', 'citywide']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Enhanced Commercial District 6', ['special_enhanced', 'citywide']],
+  ['Special St. George District', ['st_george', 'staten-island']],
+  ['Special Stapleton Waterfront District', ['stapleton', 'staten-island']],
+  ['Special South Richmond Development District', ['south_richmond', 'staten-island']],
+  ['Special Tribeca Mixed Use District', ['mixed_use', 'citywide']],
+  ['Special Downtown Brooklyn District', ['downtown_brooklyn', 'brooklyn']],
+  ['Special Ocean Parkway District', ['ocean_parkway', 'brooklyn']],
+  ['Special Enhanced Commercial District- 1', ['special_enhanced', 'citywide']],
+  ['Special Sheepshead Bay District', ['sheepshead_bay', 'brooklyn']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special College Point District', ['college_point', 'queens']],
+  ['Special Forest Hills District', ['forest_hills', 'queens']],
+  ['Special Long Island City Mixed Use District', ['long_island_city', 'queens']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Bay Ridge District', ['bay_ridge', 'brooklyn']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Mixed Use District (MX-14)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-14)', ['mixed_use', 'citywide']],
+  ['Special Enhanced Commercial District- 3', ['special_enhanced', 'citywide']],
+  ['Special Mixed Use District (MX-15)', ['mixed_use', 'citywide']],
+  ['Special Governors Island District', ['governors_island', 'manhattan']],
+  ['Special Lower Manhattan District', ['lower_manhattan', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special Transit Land Use District', ['transit_land_use', 'manhattan']],
+  ['Special United Nations Development District', ['un_development', 'manhattan']],
+  ['Special Park Improvement District', ['park', 'manhattan']],
+  ['Special Park Improvement District', ['park', 'manhattan']],
+  ['Special Mixed Use District (MX-13)', ['mixed_use', 'citywide']],
+  ['Special Planned Community Preservation District', ['planned_community', 'manhattan']],
+  ['Special Coastal Risk District', ['special_coastal_risk', 'citywide']],
+  ['Special Midtown District', ['midtown', 'manhattan']],
+  ['Special Enhanced Commercial District 5', ['special_enhanced', 'citywide']],
+  ['Special Mixed Use District (MX-16)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-16)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-16)', ['mixed_use', 'citywide']],
+  ['Special Downtown Jamaica District', ['jamaica', 'queens']],
+  ['Special Mixed Use District (MX-8)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-8)', ['mixed_use', 'citywide']],
+  ['Special Mixed Use District (MX-8)', ['mixed_use', 'citywide']],
+  ['Special United Nations Development District', ['un_development', 'manhattan']],
+  ['Special Mixed Use District (MX-16)', ['mixed_use', 'citywide']],
+  ['Special South Richmond Development District', ['south_richmond', 'staten-island']],
+  ['Special Enhanced Commercial District 5', ['special_enhanced', 'citywide']],
+  ['Special Mixed Use District (MX-16)', ['mixed_use', 'citywide']],
+  ['Special Coastal Risk District', ['special_coastal_risk', 'citywide']],
+  ['Special Downtown Far Rockaway District', ['far_rockaway', 'queens']],
+];
+
+export default specialPurposeCrosswalk;
